src/requirements.txt
google-cloud-bigquery==2.34.4
